add(feature) clear all button located at top right on notification window

+++ scripts/eww-dunst-rm.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
 
-dunstctl history-rm "$1"
+if (($# == 1)); then
+  dunstctl history-rm "$1"
+else
+  dunstctl history-clear
+fi
 $HOME/.config/eww/scripts/eww-dunst-json.sh
